Problems Page & Details Page fix: multiple files used in a grouped folder for rendering
