+++ app/contents/leetcode/page.tsx
@@ -1,12 +1,15 @@
 "use client";
 import { LeetCodeProps } from "@/app/types/types";
 import Link from "next/link";
-import { problems } from "@/data/problems";
 import { Arimo } from "next/font/google";
 import { Crimson_Text } from "next/font/google";
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import lc from "../../../public/assets/leetcode.svg";
 import Image from "next/image";
+
+// Problem data
+import { problems1 } from "@/data/problems/problems1";
+import { problems2 } from "@/data/problems/problems2";
 
 const ArimoFont = Arimo({
   subsets: ["latin"],
@@ -78,6 +81,7 @@ const Problem = ({ problem, index }: Props) => {
 const LeetCode = () => {
   const numberPerPage = 10;
   const [curr, setCurr] = useState(1);
+  const [problems, setProblems] = useState<any[]>([]);
 
   const totalPages = Math.ceil(problems.length / numberPerPage);
 
@@ -93,6 +97,14 @@ const LeetCode = () => {
   const Prev = () => {
     curr >= 2 && setCurr(curr - 1);
   };
+
+  useEffect(() => {
+    const loadProblems = () => {
+      const allProblems = [...problems1, ...problems2];
+      setProblems(allProblems);
+    };
+    loadProblems();
+  }, []);
 
   return (
     <div className="flex flex-col items-center gap-4">
